requirements.txt
opencv-python>=4.5.0
torch>=1.7.0
numpy>=1.19.0
pygame>=2.0.0
requests>=2.25.0
Pillow>=8.0.0
python-dotenv>=0.15.0
tqdm>=4.50.0 